Rework the xfce4-about code.
Use arrays to store the credit information and a separate
header for the translators that is generated by the
Transifex server.

The contributors file has been updated too. A core contributor
has to be active in the last 2 years or so in > 1 module.
Dropped the "Project Lead" group, since that is nonsense IMHO.

Also put the dialog in glade and cleanup the code.